Added check for NaN latitude and longitude values
As data is recieved, valid latitude and longitude values are stored in a cache object. If an NaN value is found, the NaN will be replaced with a previously stored value if available.

We may need to find a way to handle situations where there are no previously stored values. Alternatively, the cache can be stored in some other manner so that it is persistent.